Support bearing factors between Pose3 values
I am modelling a system in which there are two bodies, and one can
observe the other but cannot estimate the other's pose. This is
perfectly modeled by a BearingRangeFactor, but without this patch, you
cannot make a BearingRangeFactor between two Pose3 values. This adds
support for that by extending the Pose3 class to support calling
bearing() on another Pose3.