Fix LM linearizedCostChange for robust noise models
When robust noise models are used, the linearized error is not
the same as the nonlinear error even at zero delta.

In the L2 loss case, the rho(error) = 0.5 * error^2
which is the same as 0.5 *(sqrt(weight)*error)^2
since weight is always 1.0;

However this is not the case with other loss functions since in general,
rho(error) != 0.5 * weight * error^2

This PR fixes the LevenbergMarquardtOptimizer to explicityly compute
the linear cost change.